+++ example/plugins/echo.py
@@ -1,9 +1,9 @@
 from kutana import Plugin
 
 
-plugin = Plugin(name="Echo")
+plugin = Plugin(name="Echo", description="Reply with send message")
 
 
 @plugin.on_startswith_text("echo")
-async def on_echo(message, env):
+async def _(message, env):
     await env.reply("{}".format(env.body), attachment=message.attachments)