+++ example/plugins/image.py
@@ -1,11 +1,11 @@
 from kutana import Plugin, get_path
 
 
-plugin = Plugin(name="Show me")
+plugin = Plugin(name="Image", description="Send image")
 
 
-@plugin.on_text("showme")
-async def on_echo(message, env):
+@plugin.on_text("image")
+async def _(message, env):
     with open(get_path(__file__, "send_image_assets/pizza.png"), "rb") as fh:
         image = await env.upload_photo(fh.read())