+++ example/plugins/metrics.py
@@ -4,11 +4,11 @@ import time
 import os
 
 
-plugin = Plugin(name="Metrics")
+plugin = Plugin(name="Metrics", description="Send some information")
 
 
 @plugin.on_text("metrics")
-async def on_metrics(message, env):
+async def _(message, env):
     process = psutil.Process(os.getpid())
 
     taken_memory = int(process.memory_info().rss / 2**20)