+++ test/test_manager_debug.py
@@ -7,26 +7,25 @@ class TestManagerDebug(KutanaTest):
 
         with self.debug_manager(["echo message"]) as plugin:
 
-            async def on_echo(message, env):
+            @plugin.on_startswith_text("echo")
+            async def _(message, env):
                 await env.reply(env.body)
-
-            plugin.on_startswith_text("echo")(on_echo)
 
         self.assertEqual(self.manager.replies_all, {1: ["message"]})
 
     def test_request(self):
         with self.debug_manager(["echo message"]) as plugin:
 
-            async def on_echo(message, env):
+            @plugin.on_startswith_text("echo")
+            async def _(message, env):
                 self.assertEqual(await env.request("method"), None)
-
-            plugin.on_startswith_text("echo")(on_echo)
 
 
     def test_upload_photo(self):
         with self.debug_manager(["132"]) as plugin:
 
-            async def on_text(message, env):
+            @plugin.on_has_text()
+            async def _(message, env):
                 attachment = await env.upload_photo("photo")
 
                 self.assertEqual("photo", attachment)
@@ -34,16 +33,14 @@ class TestManagerDebug(KutanaTest):
                     "photo", await env.get_file_from_attachment(attachment)
                 )
 
-            plugin.on_has_text()(on_text)
-
     def test_send_message(self):
         self.target = ["message"]
 
         with self.debug_manager(["echo message"]) as plugin:
-            async def on_echo(message, env):
-                await env.send_message(env.body, 1)
 
-            plugin.on_startswith_text("echo")(on_echo)
+            @plugin.on_startswith_text("echo")
+            async def _(message, env):
+                await env.send_message(env.body, 1)
 
     def test_multiple_senders_plain(self):
         self.target = ["message1"]
@@ -55,10 +52,9 @@ class TestManagerDebug(KutanaTest):
             (3, "echo message4")
         ]) as plugin:
 
-            async def on_echo(message, env):
+            @plugin.on_startswith_text("echo")
+            async def _(message, env):
                 await env.reply(env.body)
-
-            plugin.on_startswith_text("echo")(on_echo)
 
         self.assertEqual(
             self.manager.replies_all,