+++ test/test_plugins.py
@@ -22,20 +22,18 @@ class TestPlugins(KutanaTest):
 
         with self.debug_manager(queue) as plugin:
 
-            async def on_echo(message, env):
+            @plugin.on_startswith_text("echo ", "echo\n")
+            async def _(message, env):
                 await env.reply(env.body)
-
-            plugin.on_startswith_text("echo ", "echo\n")(on_echo)
 
     def test_plugin_on_startup(self):
         self.called = 0
 
         with self.debug_manager(["message"]) as plugin:
 
+            @plugin.on_startup()
             async def startup(app):
                 self.called += 1
-
-            plugin.on_startup()(startup)
 
         self.assertEqual(self.called, 1)
 
@@ -92,10 +90,9 @@ class TestPlugins(KutanaTest):
 
         with self.debug_manager(queue) as plugin:
 
-            async def on_startswith_text(message, env):
+            @plugin.on_startswith_text("pr")
+            async def _(message, env):
                 self.assertEqual(env.args, queue_answer.pop(0))
-
-            plugin.on_startswith_text("pr")(on_startswith_text)
 
         self.assertFalse(queue_answer)
 
@@ -104,39 +101,31 @@ class TestPlugins(KutanaTest):
 
         with self.debug_manager(["123", ".hi arg1 arg2"]) as plugin:
 
-            async def on_has_text(msg, env):
+            @plugin.on_has_text("12")
+            async def _(msg, env):
                 self.assertEqual(env.found_text, "12")
-
                 await env.reply("1")
 
-            plugin.on_has_text("12")(on_has_text)
-
-
-            async def on_startswith_text(msg, env):
+            @plugin.on_startswith_text(".")
+            async def _(msg, env):
                 self.assertEqual(env.args, ["hi", "arg1", "arg2"])
                 self.assertEqual(env.body, "hi arg1 arg2")
                 self.assertEqual(env.prefix, ".")
-
                 await env.reply("2")
-
-            plugin.on_startswith_text(".")(on_startswith_text)
 
     def test_plugins_callbacks_done(self):
         self.counter = 0
 
         with self.debug_manager(["123"]) as plugin:
 
-            async def on_has_text(message, env):
+            @plugin.on_has_text("123")
+            async def _(message, env):
                 self.counter += 1
-
                 return "DONE"
 
-            plugin.on_has_text("123")(on_has_text)
-
-            async def on_has_text_2(message, env):
+            @plugin.on_has_text("123")
+            async def _(message, env):
                 self.counter += 1
-
-            plugin.on_has_text("123")(on_has_text_2)
 
         self.assertEqual(self.counter, 1)
 
@@ -145,17 +134,14 @@ class TestPlugins(KutanaTest):
 
         with self.debug_manager(["123"]) as plugin:
 
-            async def on_has_text(message, env):
+            @plugin.on_has_text("123")
+            async def _(message, env):
                 self.counter += 1
-
                 return "GOON"
 
-            plugin.on_has_text("123")(on_has_text)
-
-            async def on_has_text_2(message, env):
+            @plugin.on_has_text("123")
+            async def _(message, env):
                 self.counter += 1
-
-            plugin.on_has_text("123")(on_has_text_2)
 
         self.assertEqual(self.counter, 2)
 
@@ -167,12 +153,11 @@ class TestPlugins(KutanaTest):
             self.plugins.append(Plugin())
             self.plugins.append(Plugin())
 
-            async def on_has_text(message, env):
-                self.counter += 1
-                return "GOON"
-
             for pl in self.plugins:
-                pl.on_has_text()(on_has_text)
+                @pl.on_has_text()
+                async def _(message, env):
+                    self.counter += 1
+                    return "GOON"
 
         self.assertEqual(self.counter, 3)
 
@@ -181,16 +166,15 @@ class TestPlugins(KutanaTest):
 
         with self.debug_manager(["message"]):
 
-            async def on_has_text_early(message, env):
+            @self.plugin.on_has_text(priority=1)
+            async def _(message, env):
                 self.answers.append("early")
 
                 return "GOON"
 
-            async def on_has_text(message, env):
+            @self.plugin.on_has_text()
+            async def _(message, env):
                 self.answers.append("late")
-
-            self.plugin.on_has_text()(on_has_text)
-            self.plugin.on_has_text(priority=1)(on_has_text_early)
 
         self.assertEqual(self.answers, ["early", "late"])
 
@@ -199,16 +183,15 @@ class TestPlugins(KutanaTest):
 
         with self.debug_manager([None]):
 
-            async def on_raw_early(update, env):
+            @self.plugin.on_raw(priority=1)
+            async def _(update, env):
                 self.answers.append("early")
 
                 return "GOON"
 
-            async def on_raw(update, env):
+            @self.plugin.on_raw()
+            async def _(update, env):
                 self.answers.append("late")
-
-            self.plugin.on_raw()(on_raw)
-            self.plugin.on_raw(priority=1)(on_raw_early)
 
         self.assertEqual(self.answers, ["early", "late"])
 
@@ -217,30 +200,24 @@ class TestPlugins(KutanaTest):
         self.counter = 0
 
         with self.debug_manager(["123", "321"]) as plugin:
-            async def on_123(message, env):
+
+            @plugin.on_text("123")
+            async def _(message, env):
                 self.assertEqual(message.text, "123")
                 self.counter += 1
 
-            plugin.on_text("123")(on_123)
-
-
-            async def on_everything(message, env):
+            @plugin.on_has_text()
+            async def _(message, env):
                 self.assertNotEqual(message.text, "123")
                 self.counter += 1
 
-            plugin.on_has_text()(on_everything)
-
-
-            async def on_raw(update, env):
+            @plugin.on_raw()
+            async def _(update, env):
                 raise RuntimeError
 
-            plugin.on_raw()(on_raw)
-
-
-            async def on_dispose():
+            @plugin.on_dispose()
+            async def _():
                 self.disposed += 1
-
-            plugin.on_dispose()(on_dispose)
 
         self.assertEqual(self.counter, 2)
         self.assertEqual(self.disposed, 1)
@@ -250,18 +227,16 @@ class TestPlugins(KutanaTest):
 
         with self.debug_manager(self.target) as plugin:
 
-            async def on_echo(message, env):
+            @plugin.on_startswith_text("echo")
+            async def _(message, env):
                 await env.reply(message.text)
-
-            plugin.on_startswith_text("echo")(on_echo)
 
     def test_plugin_no_attachments_type(self):
         plugin = Plugin()
 
-        async def on_attachment(message, env):
+        @plugin.on_attachment("photo")
+        async def _(message, env):
             return "DONE"
-
-        plugin.on_attachment("photo")(on_attachment)
 
         wrapper = plugin._callbacks[0][1]
 
@@ -301,10 +276,9 @@ class TestPlugins(KutanaTest):
     def test_plugin_attachments_type(self):
         plugin = Plugin()
 
-        async def on_attachment(message, env):
+        @plugin.on_attachment("photo")
+        async def _(message, env):
             return "DONE"
-
-        plugin.on_attachment("photo")(on_attachment)
 
         wrapper = plugin._callbacks[0][1]
 
@@ -322,10 +296,9 @@ class TestPlugins(KutanaTest):
     def test_plugin_no_text(self):
         plugin = Plugin()
 
-        async def on_has_text(message, env):
+        @plugin.on_has_text()
+        async def _(message, env):
             return "DONE"
-
-        plugin.on_has_text()(on_has_text)
 
         wrapper = plugin._callbacks[0][1]
 
@@ -341,10 +314,9 @@ class TestPlugins(KutanaTest):
     def test_on_message(self):
         plugin = Plugin()
 
-        async def on_message(message, env):
+        @plugin.on_message()
+        async def _(message, env):
             return "DONE"
-
-        plugin.on_message()(on_message)
 
         wrapper = plugin._callbacks[0][1]
 
@@ -363,41 +335,29 @@ class TestPlugins(KutanaTest):
 
         with self.debug_manager(queue) as plugin:
 
-            async def no_trigger(message, env):
+            @plugin.on_attachment("photo", "audio")
+            async def _(message, env):
                 self.assertTrue(False)
 
-            plugin.on_attachment("photo", "audio")(no_trigger)
-
-
-            async def zero_trigger(message, env):
+            @plugin.on_has_text("привет")
+            async def _(message, env):
                 self.result |= 1 << 0
 
-            plugin.on_has_text("привет")(zero_trigger)
-
-
-            async def one_trigger(message, env):
+            @plugin.on_regexp_text(r"a(b|c)")
+            async def _(message, env):
                 self.assertTrue(env.match.group(0), "ab")
-
                 self.result |= 1 << 1
 
-            plugin.on_regexp_text(r"a(b|c)")(one_trigger)
-
-
-            async def two_trigger(message, env):
+            @plugin.on_regexp_text(re.compile(r"a(e|z)"))
+            async def _(message, env):
                 self.result |= 1 << 2
 
-            plugin.on_regexp_text(re.compile(r"a(e|z)"))(two_trigger)
-
-
-            async def three_trigger(message, env):
+            @plugin.on_startswith_text("hell")
+            async def _(message, env):
                 self.result |= 1 << 3
 
-            plugin.on_startswith_text("hell")(three_trigger)
-
-
-            async def four_trigger(message, env):
+            @plugin.on_text("ecHo")
+            async def _(message, env):
                 self.result |= 1 << 4
-
-            plugin.on_text("ecHo")(four_trigger)
 
         self.assertEqual(self.result, (1 << 5) - 1)